+++ src/cmsc420/schema/spatial/PRQuadTreeSeedling.java
@@ -3,7 +3,7 @@ package cmsc420.schema.spatial;
 public class PRQuadTreeSeedling implements Seedling {
 
 	@Override
-	public SpatialStructure generate(int width, int height) {
+	public SpatialStructure generate(float width, float height) {
 		return new PRQuadTree(width, height);
 	}
